backend/requirements.txt
django
requests
python-jose
openai
djangorestframework
python-dotenv
clerk-backend-api
PyJWT
plaid-python
